requirements.txt
# Core dependencies for data validation and ML pipeline
psycopg2-binary>=2.9.0
boto3>=1.28.0
PyYAML>=6.0

# Data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# ML models (for Phase D)
scikit-learn>=1.3.0
lightgbm>=4.0.0
catboost>=1.2.0
xgboost>=2.0.0  # Optional: Complete gradient boosting trio
